New simple python games has been added: number finding game and word finding one
I have uploaded two simple Python games-number a finding game and word finding one. In the number finding game, there are two parts. Firstly, computer thinks a number in a given range, and a user guesses it. Then, the reverse scenario occurs- the user thinks a number in that range, and the PC guesses it. At the end, the final result is announced: who wins the total game by the number of guesses each makes. To do so I used some random  functions like randint() and choice(). The second game, word finding, has a similar context but PC can't guess, and it is also mainly based on random method.